Update patch set 2
Patch Set 2:

"Unfortunately the build system doesn't work well if you reference C/C++ source files with "../". The workaround is to define a static library in the dir where the source code lives, and then define in cts the executable that links in the static library."

i've done this, as you can see, but i don't think it's going to work. the CTS side of things basically greps the source because for some bizarre reason it needs a list of the tests. so if my CTS thing has no source because it's just pulling in a WHOLE_STATIC_LIBRARy, we need some other way to tell it the list of tests.

Patch-set: 2
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0