Update patch set 2
Patch Set 2: Verified-1 Code-Review-2

yeah, even if this were more selective, this isn't the way we'd want to do it. we should control visibility by building libraries with a default visibility of "hidden" and explicitly annotating the functions we want to export.

in the special case of dlmalloc, exporting everything is probably a "feature", at least for the foreseeable future.

Patch-set: 2
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2
Label: Verified=-1