Update patch set 3
Patch Set 3:

Sorry for the delay, hope this addresses some of the concerns.  As far as testing goes I find it pretty difficult to test the changes I've made to the repository and downloading/compiling all of AOSP to test this feels like overkill.  All that I can confirm is a mirrored change in the branch I'm working in doesn't blow up libc. (and also that it appears to compile)

Patch-set: 3
Label: Verified=0