Update patch set 1
Patch Set 1:

I did a quick perf benchmarking on a nexus 7 and galaxy nexus and this is faster in many cases, but it's also slower in a few.

For example, the new code is slightly slower for all memcpy < 16. It's also slower for all memcpy's where the size is a multiple of 32 (by quite a bit in many cases). This seems to imply that the neon code isn't faster than the aligned code, for the smaller sizes, but that the neon code is faster at handling the copy of the tail bytes < 32. Does this match what you are seeing? As sizes increase the inefficiency in the copy goes away so that any size wins.

To do a true apples to apples comparison, I modified the code I used as a baseline to be the same as your code but without the < 128 byte check.

I need to do a bit more benchmarking to make sure that the other changes are not affecting larger sizes, but I would guess that they are very beneficial.

Patch-set: 1