Update patch set 1
Patch Set 1:

I believe the 32 size is a result of the early out in the arm only code. Specifically, the beq near the less than 32 left check, that results in the bypass of most of the other logic. I did try the neon code with the bypass but it doesn't make as much of a difference, so there is something else going on too.

How did you determine the lengths of the most used memcpy? I ask because my gut tells me that most memcpy's will be data structures that are divisible by 8, but I could be wrong. We could modify the code to optimize for these cases too.

Also, how are you benchmarking the changes? I've been using the micro_bench test, but if your results differ from mine, I want to make sure micro_bench is operating properly.

Patch-set: 1