Update patch set 4
Patch Set 4:

(really need to get back to that. iirc the point was that the existing CTS thing just basically does a grep, so it wouldn't understand repeated #inclusion or compile-time-generated names. as it turned out, it didn't cope with source not in the cts/ directory either, so we may end up writing our own, but it still sounds hard to deal with repeated #inclusion to me.)

Patch-set: 4
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0