Update patch set 1
Patch Set 1:

It doesn't match up with my results completely, but I'm using the nexus 7 and galaxy nexus devices with the cpu locked at 1.2GHz to gather data. Are you locking the cpu?

I also notice the new algorithm has much worse performance when the src pointer is 8 byte aligned and the dst pointer is 4 byte aligned. There are a couple of other corner case alignments that perform in odd ways, but I want to regather my data.

I suspect that the new code isn't prefetching as well as the old code. If your main memory is faster than my devices, that might explain why I see a larger drop off as the copy sizes increase. For example, when copying 1 million bytes or greater, the new algorithm performs at about half the speed of the current code. I'm modifying an existing benchmark to see if I can benchmark the speed of the code on uncached data to see if my supposition is correct.

Patch-set: 1