Update patch set 1
Patch Set 1:

We just made a change in cortex-strings that improves performance for the 8:4 alignment case:

=== modified file 'src/linaro-a9/memcpy.S'
--- src/linaro-a9/memcpy.S	2013-04-03 16:25:33 +0000
+++ src/linaro-a9/memcpy.S	2013-06-27 16:50:54 +0000
@@ -239,8 +239,8 @@
 .Lcpy_not_short:
 	/* At least 64 bytes to copy, but don't know the alignment yet.  */
 	str	tmp2, [sp, #-FRAME_SIZE]!
-	and	tmp2, src, #3
-	and	tmp1, dst, #3
+	and	tmp2, src, #7
+	and	tmp1, dst, #7
 	cmp	tmp1, tmp2
 	bne	.Lcpy_notaligned
 

It is expected to make more of a difference on A9 versus A15. Recently I have been working primarily with A15 and not A9, and indeed I have only benchmarked the old bionic code on the A9, so at some point I need to dig out my Pandaboard and revisit that.

I'll abandon this change and resubmit as you suggest, thanks!

Patch-set: 1