Update patch set 1
Patch Set 1: Verified+1 Code-Review+2

thanks.

do you have makefile changes to share with us for this? i'd love an extra target in the tests/ makefile that builds the bionic unit tests to run on bionic on the host (in addition to the glibc on the host target we already have), but i never have the time to get around to it...

Patch-set: 1
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2
Label: Verified=+1