Update patch set 1
Patch Set 1:

it ends up in libc.so, but some of the intermediate libraries on the way to building libc.so have clone (which is in libc_common.a) but not __thread_entry (which isn't).

i suspect the other option would be to move clone out of libc_common.a.

Patch-set: 1