Update patch set 1
Patch Set 1:

libc_common.a is also used to build .so files, other
than libc.so.  We can either use this patch, which
uses PLT for __thread_entry.  When building libc.so,
linker will skip PLT since it is defined in libc.so.
For other .so files, linker will keep PLT.  Or we
can move thread_create.o to libc_common.a.

Patch-set: 1