Update patch set 1
Patch Set 1: Verified+1

this is statvfs(2), not stat(2). the flags are file system flags, and sadly the flags on the /proc file system vary. most of the other fields are pretty unpredictable too, but checking the first and last fields (which are luckily the most predictable) seems like a good way to be reasonably confident about the fields in between.

tests pass on z620 host and ARM device.

Patch-set: 1
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: SUBM=+1
Label: Verified=+1