Update patch set 3
Patch Set 3:

I looked and yeah that bug does appear to be fixed on my machine too. I've tried every combination of parameter to clone, but simply making the call causes things to get messed up later for pthread. I searched and searched for any mention of something similar to what I saw or some way that we are using clone improperly, and nothing seems to match. I even looked to see if mixing clone and pthread is bad, but I couldn't find anything.

Since I don't have a bug, and it might not be a glibc bug, I've changed the comment slightly, but it's probably not worth spending any more time on this. I can turn the test off completely for glibc if that's the better solution.

Patch-set: 3
Label: Verified=0