Update patch set 4
Patch Set 4: Code-Review+2

i certainly don't like that the test is _different_ for the two libraries. just not running it at all and admitting that we don't know why it doesn't work might be better.

i wonder if it's because glibc's pthread implementation uses the extra clone flags to have the kernel update thread state? maybe when our clone exits the kernel clobbers some state for the main thread because we didn't unlink ourselves somehow?

Patch-set: 4
Label: Code-Review=+2