Update patch set 2
Patch Set 2:

It isn't. Posix says you should return 1.0 and the 2013 edition is pretty unambiguous about it.

"If x or y is a NaN, a NaN shall be returned (unless specified elsewhere in this description). <... description continues ...> For any value of y (including NaN), if x is +1, 1.0 shall be returned"

pow is specified to set errno to EDOM only when x < 0 and y is finite and non integral.

Patch-set: 2