Update patch set 1
Patch Set 1:

I think it should be abandoned. I have on my plate to get the new linaro version of memcpy for cortex-a9 up to snuff. That version is definitely faster in most cases than this version.  The linaro version only needs to be tuned for the few cases that are not optimized.

Patch-set: 1
Reviewer: Christopher Ferris <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0