Update patch set 1
Patch Set 1:

These are the times I wish C supported some way to undeclare a function.... We could define it, use it in the inlined function, then remove it.

I don't have any better solution. If others are happy with the current patch, then I'm fine with it too. It's not *that* big of a deal publicizing the functions...

Patch-set: 1