Update patch set 1
Patch Set 1:

I see. I don't know the structure of Bionic, but I'd have thought that an alias to memset/memcpy/memmove would be a lot simpler than this. Anyway, though it's quite ugly, this seem to fix the issue, at least temporarily.

I'm not sure why the androideabi behaves like aeabi in Clang. In the long term, if Bionic is more similar to glibc, than we shouldn't change the names of the library calls for the androideabi in the first place.

Patch-set: 1