Update patch set 1
Patch Set 1:

What is this needed for? I ask because there are a couple of places where these same values are defined in other code, but the values are different. For example, art has them, and libunwind has some. They both use slightly different values, but I don't know if it will cause any conflicts.

Patch-set: 1