Update patch set 1
Patch Set 1:

This is not to put the 64bit and 32bit libs in the same directory.

This is for the LD_LIBRARY_PATH use case. When LD_LIBRARY_PATH is used, the linker could seek the 32bit and 64bit library in the same directory. If the LD_LIBRARY_PATH is defined as:
   xxxx/lib:xxxx/lib64
the libs in xxxx/lib could be always tried first. The xxxx/lib64 is never tried if the libs has the same name for 32bit and 64bit.

Define it as xxxx/lib64:xxxx/lib has similar issue.

Patch-set: 1
Reviewer: Fengwei Yin <1011068@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0