Update patch set 7
Patch Set 7:

That would require that every time we add a test, someone would have to regenerate the list. In addition, we would still require most of these changes because I don't think cts supports a different list of tests per platform. The only savings would be that we don't need emtpy tests for glibc only stuff (which I don't think we have a lot of anyway).

This is a case of the least bad of the bad choices.

Patch-set: 7
Label: Verified=0