Update patch set 1
Patch Set 1:

> If we changed the parsers to rename _NSIG, is there any code that
 > depends on it having the current value? Maybe something in
 > external?

there's lots of code that refers to NSIG and _NSIG. the real question is what number they think they're getting :-)

glibc has NSIG == _NSIG if that's what you're asking. so it seems unlikely anyone is actually expecting the kernel's value.

our old value was doubly wrong.
1. it wasn't the "+1" userspace value it should be.
2. it wasn't the right kernel value either (32 rather than 64) except on mips.

looking at the users, there's a mix. there's code that's clearly assuming +1 and is currently incorrect, and of the code that wouldn't be right if we fixed things (mksh, say) most of that is already wrong; just a question of off by 1 or off by 2 :-)

Patch-set: 1
Label: Verified=0