Update patch set 1
Patch Set 1:

I believe this should work for arm64, and probably x86_64, not familiar enough with mips64 to know if it will cause problems there.  It may not be completely optimal for x86_64 if movl $1, %eax is more efficient than movl $1, %rax.

An alternate solution is to add a __set_errno variant that returns long for use by syscalls that return 64-bit types.

Patch-set: 1
Label: Verified=0