Update patch set 1
Patch Set 1:

I know that the indentation is messed up here. In fact, it's messed up in various places unrelated to the updated... so when I tried to fix I introduced lots of unrelated changes.

To make it easier for review I left it with broken indentation and upload another CL which fix the indentation.

Ideally I would like to patch this CL with the indentation fixes after +2 instead of having a separate one.

Patch-set: 1