Update patch set 1
Patch Set 1:

I was 'expanding' the propagation of this flag to all. If we need to refine to libbionic.a (libbionic.so? have not checked), how would you recommend doing so to be more precise? Do we really want to be more precise?

We will *not* be removing this, per se, we will change it's tense to default to enabling LOGD (change to ! TARGET_USES_KERNEL_LOGGER <sigh> wish I had thought about 'phase 2' more carefully)

Patch-set: 1