Update patch set 1
Patch Set 1:

A recursive locking has a cost in checks and other added operations; this implementation remains viable because it is still fast, adding a new burden to it for demangling in not worth the feature (in my opinion).

Changing libgcc-demangle is feasible but I don't like much the idea of changing a separate project.

This solution seems appropriate to me.

Patch-set: 1