Update patch set 1
Patch Set 1: Verified+1

it seems to work (running the static version of the unit tests), and we're relying on this already, but i agree that this seems fishy. it looks like GCC supports constructor priorities so i'll give the libc tls-setting constructor one priority and all the others a higher priority in a later patch. (the dynamic linker will just go through the constructors in the order they appear in the array in the ELF file, so we don't need to change anything there.)

Patch-set: 1
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1