Update patch set 3
Patch Set 3:

PS3 just fixes a comment.

I'm not sure how to do that, neither that we should do it. Different 64bit arches may support a different number and not necessary > 32.

From what I can tell x86_x64 kernel supports 64 CPUs (at least based on its config). Note that I used the same as glibc here which might not be necessary the best option but it's definitely the most flexible one. (On the side, my guess is that glibc choose the value because of powerpc whose kernel config indicate 1024 CPUs)

Patch-set: 3
Label: Verified=0