Update patch set 3
Patch Set 3:

> Is there any way to add a test to verify that 64 bit arches can
 > support > 32 cpus?

i think it depends what you mean. if you mean "does the kernel support a large cpuset size?", we can easily just pass a large one. looking at the source a kernel will get grumpy if the set size is too small or not a multiple of "sizeof(unsigned long)" bytes, but it can be arbitrarily large (until you ENOMEM inside the kernel). so we could easily add a test for 8192 or something.

if you mean "does the kernel support running on a large number of cpus?" i'm (a) not sure how we'd test that and (b) don't think such a test is related to this API.

so i'm assuming you mean the former, and agree that testing with a large number (which we can ASSERT_GT our default size) would make a better test.

Patch-set: 3
Label: Verified=0