Update patch set 1
Patch Set 1:

Unfortunately, this makes the test program entirely crash on arm64, because of the toolchain bug with virtual inheritance mentioned in 14265681, so we probably shouldn't land this until the toolchain is fixed. The case that was breaking on 64-bit (library with no relro section) is now explicitly covered by the test, though, and passes on 32-bit locally.

Patch-set: 1