Update patch set 1
Patch Set 1:

The same can be observed on linux. I don't understand how it's related to this patch. My purpose here is to make linker a bit more compatible with ld.so to handle some legacy code. 
Do you have any test case which is going to be broken by this patch?

Patch-set: 1