Update patch set 2
Patch Set 2:

The new test doesn't appear to catch any failure with the previous version of the memmove function. Is this performance fix or a correctness fix?

I believe that another test overlaps with this new test which is why I would be surprised that it would fail, while the other would not.

Patch-set: 2
Label: Code-Review=0