Update patch set 1
Patch Set 1:

It seems to me that the likely core issue here is that if a key is deleted and reallocated, we need to clear corresponding per-thread data for all threads, so that values for the new key appear to be NULL?  That seems tricky.  I didn't check whether the old code is correct.

I think there is also a race if a thread uses a newly created key and exits before the destructor is installed.

The performance issues seem to have been anticipated by the Posix spec, in that they discourage checking key validity, esp. in pthread_getspecific().  The old validity check did not enforce memory ordering, and was thus not too expensive, though probably still undesirable.  The new one enforces memory ordering, making it much worse.  That could be fixed by either dropping the expensive part of the check or making it memory_order_relaxed.

I'm not sure I would go in this direction at all.  But I don't think the observed performance problem is the main reason for that.  If there is a strong reason to do this, I would merge the destructor and in_use bit, using a special value for (no destructor but in use), which may also be useful while clearing data across all threads.

Patch-set: 1
Reviewer: Hans Boehm <1042828@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0