Update patch set 1
Patch Set 1:

> Which means if we define it as byte on IA, we don't need atomic operation on reading/updating itself.

I'm guessing 'atomic' is there for the memory ordering it provides.

> it's highly unlikely any code is (a) recycling keys and (b) taking advantage of the fact that they're supposed to be initialized to NULL

It is, however, pretty likely that some code creates a key and assumes that the values will be NULL ("get; if NULL: create"). If someone (not necessarily the same code) deletes a key before that, you'll take over their old key (keys are provided lowest-free-first) and have a mess.

> [from the issue discussion:] right now the poor performance of pthread_getspecific is the blocker

Maybe using a weaker ordering would improve performance? Looks like the default is memory_order_seq_cst, which sounds very expensive. I think we should be able to get away with memory_order_acq_rel.

Not sure if it will help, but it's simple enough to be worth a quick test, right? :)

Using memory_order_consume for the load is probably even cheaper, but it would require the actual map values to be atomics as well (since the valid load wouldn't make the latest map contents observable). So, which is faster: one acquire or two consumes?

Note that we can skip the validity check in getspecific() altogether:

> The effect of calling pthread_getspecific() or pthread_setspecific() with a key value not obtained from pthread_key_create() or after key has been deleted with pthread_key_delete() is undefined.

But we'd still need those atomic map values to make sure that all threads see NULL after key delete + recreate (like Hans said), so I don't think that really helps. Plus, it's nice if the lib can tell us when we mess things up.

Patch-set: 1
Reviewer: Snild Dolkow <1041495@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0