Update patch set 1
Patch Set 1:

> (1 comment)

Oh, sorry, this is part of it too.  It fails when a name starts or ends with ".", if there are 2 periods in a row, or if the name is empty.  I wrote a fix for those cases and discovered that it then begins to fail when there are too many added (around 850 properties seems to fill it).

Patch-set: 1
Reviewer: Brigid Smith <1047249@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0