Update patch set 2
Patch Set 2:

> sgtm. Wasn't sure which level we wanted to fix it at. I'll make the
 > change.

the reason i meant to include a free '\n' with every call was compatibility with regular Android logging. you may see \n at the end of the strings people log, but they're superfluous. one will always be added for you.

(the other reason is that we don't want to have to find every single caller of __libc_fatal :-) )

Patch-set: 2
Label: Verified=0