Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

Script generates excessive symbols for x86 (I don't know about arm), but I'm not sure that hard coding of missing symbols in script is a better solution than mine. Actually, I've checked (for silvermont and haswell) that only popcount symbols are missing and it has an explanation as starting with slm popcount is supported in HW. That is why I believe that for now having a two-symbols list is better, at least it's crystal clear.

Patch-set: 6