Update patch set 2
Patch Set 2:

I thought of shim because most of the functions are not doing anything but avoiding malloc calls. I'll figure out a better name and add better comments.

Yes, it's not really possible to put this in libunwind since it's using the global variable. In addition, the other debug_XX code uses some of these functions and has to be avoided.

Patch-set: 2
Label: Verified=0