Update patch set 1
Patch Set 1:

i agree, its not a standard use for android, more a test of the recent changes in aosp/master with busybox.

Example, when we call "login shell" as root, sulogin or getty, a new environment is set for the new user. In busybox, some env variables are copied from the original session, like PATH, by calling setenv. This cause the libc to segfault because the new environment** var is null. Maybe a special function exists to init this var in libc, for the moment i just made a malloc of a pointer, if the var is null :

    environ = (char **)xzalloc(sizeof(char *));

Patch-set: 1
Reviewer: Tanguy Pruvot <1006481@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0