Update patch set 2
Patch Set 2:

> can you expand a bit on the reasons for reverting this? :)

We had originally removed it because the two are mostly the same (doing anything but calling exec() after vfork() is undefined behavior), but fork() will do a lot of cleanup that vfork() doesn't do. Since getting the true vfork() behavior back would require the user to write assembly for each arch, we decided to just put it back.

Patch-set: 2