Update patch set 1
Patch Set 1:

I changed the test so that it actually fails on HH with some reasonable probability if I use memory_order_relaxed everywhere.  With that (temporary) change, it still fails surprisingly rarely on HH.  Even more surprisingly, when it does fail the ordering test, it seems way off, not just by 1 or 2, as I would have expected.  I wonder whether something is going wrong with the loads for printing the error message.

Perhaps of more short term interest: Disassembling the output shows that we are currently using "dmb sy" as the ARM memory fence everywhere.  My understanding is that we should be using "dmb ish", which is claimed to be appreciably faster, though that probably varies.  There was a proposed, but incorrect gcc patch to fix this in March, but it looks like nothing is in our tool chain.  I don't know if there is a practical way to fix this.

Patch-set: 1